refactor(discovery): extract migration logic into lib/migration.ts (#721)
discovery.ts had grown to 1300+ lines mixing read-only system
inspection (discoverSystemdServices) with state-mutating migration
(migrateService, mergeServices) and a deep backup/rollback machinery.
This change splits them so reviewers can reason about either concern
in isolation, and so the diagnose / probe stack can import discovery
without dragging in backup-archive + tar / yaml / quadlet directive
handling.

- lib/discovery.ts (new size: ~290 lines): pure inspection plus
  deleteBundleResources (which is bundle-purge, the inverse of
  discovery, sharing the same primitive set).
- lib/migration.ts (new, ~960 lines): MigrationPlan, MergeOptions,
  PlanValidation, PlanFileMapping types; the createBackup /
  createBackupArchive / rollbackManagedService / stopLegacyServices /
  recordMigrationHistory helpers; analyze + preview + plan +
  execute pipelines; and the two exported entry points
  migrateService and mergeServices.

DiscoveredService stays exported from discovery and re-exported from
migration so existing call sites keep working with a one-line import
swap; the two affected routes (discovery/migrate, discovery/merge)
now import migrateService / mergeServices from @/lib/migration.

Tests: full suite (1113 passing) unchanged.
